requirements.txt
requests==2.25.1
Pillow==8.2.0
beautifulsoup4==4.9.3
colorama==0.4.3